Added StepAfterNextRootNode Strategy
This stepping strategy will run to the next root node, remember it,
and execute all subexpressions. Once it returns and arrives on the
return edge, suspend execution.

StepAfterNextRootNode needs to use source sections for equality
comparison. Because of node specializations, we can't use node identity
to check whether the correct root node was hit again. Instead, we need
to use the source sections. At this point, we are already sure we are
at a root node, and, it is the only other root node in the system, that
is going to have the same source section.

Signed-off-by: Stefan Marr <git@stefan-marr.de>
Co-authored-by: carmen <carmentorreslopez87@gmail.com>