StepAfterNextRootNode needs to use source sections for equality comparison
Because of node specializations, we can't use node identity to check
whether the correct root node was hit again. Instead, we need to use the
source sections. At this point, we are already sure we are at a root
node, and, it is the only other root node in the system, that is going
to have the same source section.