Implement low verbosity level for mx profdiff.
Low verbosity level removes detailed phases, sorts phases where order is unimportant,
and sorts the inlining tree - so that changes caused by a different order are not
displayed. Tree operations to support this were implemented. Implemented the DeltaTree,
which is a different view of the edit script that is better suited for traversal
and modifications.